improved network login flow

+++ migrations/postgres/init.sql
@@ -1,6 +1,6 @@
 -- Create players table
 CREATE TABLE IF NOT EXISTS players (
-    player_id int PRIMARY KEY,
+    player_id bigint PRIMARY KEY,
     timestamp bigint NOT NULL,
     x float NOT NULL,
     y float NOT NULL
